test(api): one api_path per contract test file
`assert_api_response` resolves a call to the first declared api_path whose
template has a path param, so sibling paths sharing a class answered for
each other: /projects/{id} and /projects/{id}/archive only stayed correct
because their verbs happened to differ, and a third PUT would have picked
the wrong one silently.

Give every api_path its own file and class. The rule now holds across the
whole directory (43 files, 43 paths) instead of being rediscovered per
domain, so the per-file comments explaining it move to openapi_helper.rb.

Same 284 runs / 930 assertions as before, and the generated schema is
byte-identical.
