+++ test/integration/api/v1/projects_test.rb
@@ -62,115 +62,12 @@ module Api
         end
       end
 
-      api_path "/projects/{id}" do
-        parameter name: "id", in: :path, schema: {type: :string, format: :uuid}, required: true
-
-        get("Get Project") do
-          operationId "project"
-          tags "Projects"
-          produces "application/json"
-
-          response(200, "successful") do
-            schema ::V1::Schemas::Project
-          end
-
-          response(404, "not found") do
-            schema ::V1::Schemas::StandardError
-          end
-
-          response(401, "unauthorized") do
-            schema ::V1::Schemas::StandardError
-          end
-        end
-
-        patch("Update Project") do
-          operationId "updateProject"
-          tags "Projects"
-          consumes "application/json"
-          produces "application/json"
-
-          request_body required: true, content: {
-            "application/json" => {schema: ::V1::Schemas::Inputs::ProjectInput}
-          }
-
-          response(200, "successful") do
-            schema ::V1::Schemas::Project
-          end
-
-          response(400, "bad request") do
-            schema ::V1::Schemas::ValidationError
-          end
-
-          response(404, "not found") do
-            schema ::V1::Schemas::StandardError
-          end
-
-          response(401, "unauthorized") do
-            schema ::V1::Schemas::StandardError
-          end
-        end
-
-        delete("Destroy Project") do
-          operationId "destroyProject"
-          tags "Projects"
-          produces "application/json"
-
-          response(200, "successful") do
-            schema ::V1::Schemas::Message
-          end
-
-          response(400, "bad request") do
-            schema ::V1::Schemas::ValidationError
-          end
-
-          response(404, "not found") do
-            schema ::V1::Schemas::StandardError
-          end
-
-          response(401, "unauthorized") do
-            schema ::V1::Schemas::StandardError
-          end
-        end
-      end
-
-      api_path "/projects/{id}/archive" do
-        parameter name: "id", in: :path, schema: {type: :string, format: :uuid}, required: true
-
-        put("Archive Project") do
-          operationId "archiveProject"
-          tags "Projects"
-          produces "application/json"
-
-          response(200, "successful") do
-            schema ::V1::Schemas::Message
-          end
-
-          response(400, "already archived") do
-            schema ::V1::Schemas::ValidationError
-          end
-
-          response(404, "not found") do
-            schema ::V1::Schemas::StandardError
-          end
-
-          response(401, "unauthorized") do
-            schema ::V1::Schemas::StandardError
-          end
-        end
-      end
-
       let(:data) { users :data }
       let(:project) { projects :narendra3 }
 
       describe "unauthorized" do
         it "does not list projects" do
           assert_api_response :get, 401
-        end
-
-        it "does not destroy a project" do
-          assert_api_response :delete, 401, path_params: {id: project.id}
-
-          assert Project.exists?(project.id)
         end
       end
 
@@ -189,16 +86,6 @@ module Api
         it "excludes ids passed in withoutIds" do
           assert_api_response :get, 200, params: {withoutIds: [project.id]} do
             refute_includes parsed_body.map { |item| item["id"] }, project.id
-          end
-        end
-
-        it "is not found for a project from another account" do
-          assert_api_response :delete, 404, path_params: {id: SecureRandom.uuid}
-        end
-
-        it "shows a project" do
-          assert_api_response :get, 200, path_params: {id: project.id} do
-            assert_equal project.id, parsed_body["id"]
           end
         end
 
@@ -227,83 +114,6 @@ module Api
           assert_no_difference "Project.count" do
             assert_api_response :post, 403, body: {name: "Infiltration", customer_id: customers(:defiant_customer).id}
           end
-        end
-
-        it "updates a project" do
-          assert_api_response :patch, 200, path_params: {id: project.id}, body: {name: "Narendra III"} do
-            assert_equal "Narendra III", parsed_body["name"]
-          end
-
-          assert_equal "Narendra III", project.reload.name
-        end
-
-        it "archives a project" do
-          assert_api_response :put, 200, path_params: {id: project.id}
-
-          assert project.reload.archived?
-        end
-
-        it "refuses to archive twice" do
-          project.archive!
-          project.save
-
-          assert_api_response :put, 400, path_params: {id: project.id} do
-            assert_equal "validation_error.project.archive", parsed_body["code"]
-          end
-        end
-      end
-    end
-
-    # Own class: /projects/{id}/archive and /unarchive are both PUT with a path
-    # param, and assert_api_response resolves to the first match.
-    class ProjectUnarchiveTest < ActionDispatch::IntegrationTest
-      include OpenapiRuby::Adapters::Minitest::DSL
-
-      openapi_schema :"v1/schema"
-
-      api_path "/projects/{id}/unarchive" do
-        parameter name: "id", in: :path, schema: {type: :string, format: :uuid}, required: true
-
-        put("Unarchive Project") do
-          operationId "unarchiveProject"
-          tags "Projects"
-          produces "application/json"
-
-          response(200, "successful") do
-            schema ::V1::Schemas::Message
-          end
-
-          response(400, "not archived") do
-            schema ::V1::Schemas::ValidationError
-          end
-
-          response(404, "not found") do
-            schema ::V1::Schemas::StandardError
-          end
-
-          response(401, "unauthorized") do
-            schema ::V1::Schemas::StandardError
-          end
-        end
-      end
-
-      let(:data) { users :data }
-      let(:project) { projects :narendra3 }
-
-      before { sign_in data }
-
-      it "unarchives an archived project" do
-        project.archive!
-        project.save
-
-        assert_api_response :put, 200, path_params: {id: project.id}
-
-        refute project.reload.archived?
-      end
-
-      it "refuses to unarchive an active project" do
-        assert_api_response :put, 400, path_params: {id: project.id} do
-          assert_equal "validation_error.project.unarchive", parsed_body["code"]
         end
       end
     end
